requirements.txt
Django==3.2.7
Pillow==8.3.2
djangorestframework==3.12.4
binance-connector==1.5.0
